Set up React Router and Redux -- Incomplete Store Page
