+++ src/pages/home.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react'
+
+class StoreFront extends Component {
+    constructor(props) {
+        super()
+        console.log(this.props)
+    }
+
+    render () {
+        return (
+            <h1>Building Up -- Home Page</h1>
+        )
+    }
+}
+
+export default StoreFront
