+++ src/pages/storefront.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react'
+
+class StoreFront extends Component {
+    constructor(props) {
+        super()
+        console.log(this.props)
+    }
+
+    render () {
+        return (
+            <h1>This is the store page</h1>
+        )
+    }
+}
+
+export default StoreFront
+
