+++ src/shared/App.js
@@ -1,0 +1,32 @@
+import React from 'react';
+import { ApolloClient, InMemoryCache, ApolloProvider } from "@apollo/client";
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+} from "react-router-dom";
+import StoreFront from '../pages/storefront'
+import Home from '../pages/home'
+
+
+const client = new ApolloClient({
+  uri: "http://localhost:4000/",
+  cache: new InMemoryCache()
+});
+
+function App() {
+  return (
+    <ApolloProvider client={client}>
+      <Router>
+        <React.Fragment>
+          <Switch>
+            <Route path="/store/:id" component={StoreFront}/>
+            <Route path="/" component={Home} />
+          </Switch>
+        </React.Fragment>
+      </Router>
+    </ApolloProvider>
+  );
+}
+
+export default App;
